dev *filter from hess-ems, *deadzone from hess-ems, +props_to_para_dict
The source code now utilises the lowpass() function and deadzone()
function from the HESS-EMS package for the FilterController and
DeadzoneController() implementation.

The abstract SimComponent class in simulation.py now features a
props_to_para_dict() method, to effectively convert data formats from
the HESS-EMS-SIM package to the HESS-EMS package.

The *Controller classes now default parameters to None in __init__, so
the parameters default to the ones defined in the HESS-EMS package if
not explicitely set.

On branch develop
Changes to be committed:
      modified:   hessemssim/__init__.py
      modified:   hessemssim/controllers.py
      modified:   hessemssim/simulation.py
      modified:   hessemssim/util.py