example +ems_comparison that compares all ems for two examples
Made a file that simulates all EMS for two random signals and plots
them in paper-quality in a joint plot.

Added extra keywords and logic in the
simulation.SimSetup.plot_timeseries() plot function, as this one
is utilized by the plot_comparison() function and this eases
customization of the plots.

Non-Implemented controllers in controllers.py (these are
FuzzyController, MPCController, NeuralController) now act as a
PassthroughController as dummy/mockup until implemented, so they can be
used in the plot_comparison function.

On branch develop
Changes to be committed:
      new file:   examples/ems_comparison.py
      modified:   hessemssim/__init__.py
      modified:   hessemssim/controllers.py
      modified:   hessemssim/simulation.py