dev *storage has now init parameter
The unified storage constructor now has a parameter `init` in [0, 1]
that defines the initial energy content as state of energy.

The InputData class now supports adding a name for the input data. If
not set, an arbitrary name based on the md5 hash is generated.

On branch develop
Changes to be committed:
      modified:   hessemssim/simulation.py
      modified:   hessemssim/storages.py